OnMarkLeaveEvent 變更為獨立 class
